requirements.txt
Flask
mysql-connector-python
configparser
requests
pandas